+++ src/components/SpeakerSelector/styles.js
